# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sanber Car API",
        "description": "sanber-car final project API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get orders",
                "description": "Get orders",
                "operationId": "index_orders",
                "responses": {
                    "200": {
                        "description": "Successful get data orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Create order",
                "description": "Create new order",
                "operationId": "store_orders",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "car_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation Error"
                    },
                    "200": {
                        "description": "Successful add data types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/user/{status}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get all ordered cars based on payment_status",
                "description": "Get all ordered cars based on payment_status",
                "operationId": "getOrderedCarsBasedOnPaymentStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Parameter payment_status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "pending": {
                                "summary": "Parameter payment status (complete|pending|error|closed)",
                                "value": "pending"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get data ordered cars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/seller": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get all ordered cars based on current seller",
                "description": "Get all ordered cars based on current seller",
                "operationId": "getOrderedCarsBasedOnSeller",
                "responses": {
                    "200": {
                        "description": "Successful get data ordered cars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get order detail",
                "description": "Get order detail",
                "operationId": "show_orders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Parameter id.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get data order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "orders"
                ],
                "summary": "Edit orders",
                "description": "Edit orders",
                "operationId": "update_orders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Parameter id.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "payment_status": {
                                        "type": "string"
                                    },
                                    "payment_url": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "integer"
                                    },
                                    "car_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation Error"
                    },
                    "200": {
                        "description": "Successful add data types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "summary": "Delete orders",
                "description": "Delete orders",
                "operationId": "destroy_orders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Parameter id.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful delete data order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/checkout/{id}": {
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Checkout order",
                "description": "Checkout order",
                "operationId": "checkout_orders",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Parameter id.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful checkout order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}